Changed types of all values containing money from Double to Int #51

+++ src/main/scala/spoker/betting/Action.scala
@@ -3,7 +3,7 @@ package spoker.betting
 sealed trait Action
 
 trait BettingAction {
-  val value: Double
+  val value: Int
 }
 
 object Call extends Action
@@ -14,6 +14,6 @@ object Fold extends Action
 
 object AllIn extends Action
 
-case class Raise(value: Double) extends Action with BettingAction
+case class Raise(value: Int) extends Action with BettingAction
 
-case class Bet(value: Double) extends Action with BettingAction
+case class Bet(value: Int) extends Action with BettingAction
